pyproject.toml
[tool.black]
line-length = 99
include = '/project$|\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.eggs
  | \.mypy_cache
  | \.tox
  | \.venv
  | \.vscode
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"

[tool.mypy]
follow_imports = "silent"
warn_redundant_casts = true
warn_unused_ignores = true
disallow_any_generics = true
check_untyped_defs = true
no_implicit_reexport = true

[tool.poetry]
authors = ["Padraic Slattery <pgoslatara@gmail.com>"]
classifiers = [
  "Programming Language :: Python",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
]
description = "Version control your dbt Cloud jobs with YML."
keywords = [
  "python",
  "cli",
  "dbt",
  "CI/CD",
]
license ="MIT"
maintainers = ["Padraic Slattery <pgoslatara@gmail.com>"]
name = "dbt-cloud-jobs"
packages = [
    { include = "dbt_cloud_jobs" },
]
readme = "README.md"
repository = "https://github.com/pgoslatara/dbt-cloud-jobs"
version = "0.0.0"

[tool.poetry.dependencies]
python = ">=3.8,<3.13"
requests = "*"
pydantic = "<2"
pyyaml = "*"

[tool.poetry.dev-dependencies]
pre-commit = "^3"
pytest = "*"
pytest-xdist = "*"

[tool.poetry.group.test.dependencies]
pytest-cov = "*"

[tool.poetry.scripts]
dbt_cloud_jobs = "dbt_cloud_jobs.main:main"

[tool.poetry_bumpversion.file."./dbt_cloud_jobs/version.py"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
